Fix: Divide Diary API serializers into multiple files
